examples/agentic/requirements.txt
langchain==0.3.1
langchain_community==0.3.1
ibm-generative-ai==3.0.0
langchain_ollama==0.2.0
langchain-ibm==0.2.1
langchain-openai==0.2.14
langgraph==0.2.35
ibm_watsonx_ai==1.1.11
python-dotenv
pypdf
replicate
black==24.10.0
